player jump fix
fixed player jumping higher after time passing

+++ Jumpy/Player.h
@@ -8,9 +8,9 @@ public:
 	bool isGrounded;
 	bool isJumping;
 
-	float moveSpeed;
+	int moveSpeed;
 	float velocityY;
-	float gravity;
+	int gravity;
 
 	int score;
